Refactor out user menu to it's own component with menu slot
